refactor(recording): promote the recording package to internal/recording
The per-request recorder is a high-level, request-scoped observability
component — the flag system decides whether it records, protocol code only
reports into it, and (in upcoming work) the client transport layer reads it
too. Living under internal/protocolserver/ misstated that role as a
protocol-layer detail. Pure mechanical move: git mv plus import-path
rewrite, no code or behavior change.

Its test-helper package internal/server/recordingtest moves alongside it
to internal/recording/recordingtest, for the same reason — the helpers
exist to test the recording package, not internal/server.
