+++ internal/protocolserver/anthropic_recording_e2e_test.go
@@ -11,7 +11,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"github.com/tingly-dev/tingly-box/internal/obs"
-	"github.com/tingly-dev/tingly-box/internal/server/recordingtest"
+	"github.com/tingly-dev/tingly-box/internal/recording/recordingtest"
 	"github.com/tingly-dev/tingly-box/internal/typ"
 )
