+++ internal/protocolserver/protocol_dispatch.go
@@ -20,7 +20,7 @@ import (
 	"github.com/tingly-dev/tingly-box/internal/protocol/transform"
 	usagepkg "github.com/tingly-dev/tingly-box/internal/protocol/usage"
 	"github.com/tingly-dev/tingly-box/internal/protocolserver/forwarding"
-	"github.com/tingly-dev/tingly-box/internal/protocolserver/recording"
+	"github.com/tingly-dev/tingly-box/internal/recording"
 	"github.com/tingly-dev/tingly-box/internal/server/config"
 	"github.com/tingly-dev/tingly-box/internal/typ"
 )
