+++ internal/protocolserver/protocol_handler.go
@@ -25,8 +25,8 @@ import (
 	"github.com/tingly-dev/tingly-box/internal/obs"
 	pkgotel "github.com/tingly-dev/tingly-box/internal/otel"
 	"github.com/tingly-dev/tingly-box/internal/otel/tracker"
-	"github.com/tingly-dev/tingly-box/internal/protocolserver/recording"
 	"github.com/tingly-dev/tingly-box/internal/protocolserver/servertool"
+	"github.com/tingly-dev/tingly-box/internal/recording"
 	"github.com/tingly-dev/tingly-box/internal/routing"
 	"github.com/tingly-dev/tingly-box/internal/server/config"
 	"github.com/tingly-dev/tingly-box/internal/typ"
