+++ internal/recording/hook.go
