+++ internal/recording/recordingtest/helpers.go
@@ -1,10 +1,10 @@
 // Package recordingtest provides shared test helpers for exercising
-// internal/protocolserver/recording's AttachRecorderHooks/ProtocolRecorder
+// internal/recording's AttachRecorderHooks/ProtocolRecorder
 // wiring through the production *protocolserver.ProtocolHandler entry points.
 //
 // This is a plain (non-_test.go) package rather than an external
 // recording_test file because its helpers must be importable both from
-// internal/protocolserver/recording's own tests and from internal/server's
+// internal/recording's own tests and from internal/server's
 // e2e tests — Go test files, even in an external _test package, are never
 // importable from another package. Putting the helpers in a normal package
 // that imports internal/protocolserver avoids that restriction; it stays
